# Conflicts:
#	ccv/chains/evm/gobindings/generated/latest/cctp_verifier/cctp_verifier.go
#	ccv/chains/evm/gobindings/generated/latest/committee_verifier/committee_verifier.go
#	ccv/chains/evm/gobindings/generated/latest/lombard_verifier/lombard_verifier.go
#	chains/evm/.gas-snapshot
#	chains/evm/gobindings/generation/generated-wrapper-dependency-versions-do-not-edit.txt

+++ deployment/v1_7_0/adapters/lombard.go
@@ -64,10 +64,10 @@ type RemoteLombardChainConfig struct {
 	LombardChainId         uint32
 	// RemoteAdapter is the optional remote token identifier accepted by Lombard bridge.
 	// Accepts either an EVM address hex string or a 32-byte hex string. Leave empty to disable adapter override.
-	RemoteAdapter      string
+	RemoteAdapter string
 	FeeUSDCents        uint16
 	GasForVerification uint32
-	PayloadSizeBytes   uint32
+	PayloadSizeBytes      uint16
 }
 
 // ConfigureLombardChainForLanesDeps are the dependencies for the ConfigureLombardChainForLanes sequence.
